requirements.txt
fastapi
psycopg2-binary
sentence-transformers
transformers
datasets
peft
tensorflow
openai
uvicorn